Shader update
Changed Shader into a super class, to be used as a basis for shaders.
--> no static
The vertex and fragment shaders will never be used again, so we don`t
need to store their ids.
The uniform code moved into the standard shader StdShader.java.
Added bind, unbind, destroy methods.

The StdShader class now contains the standard shader and the uniform
locations.

Shader calls moved from the renderable to the Masterrenderer, and
because of that the transform of the renderable must be readable.